+++ axelor-contract/src/main/java/com/axelor/apps/contract/service/ContractServiceImpl.java
@@ -492,14 +492,17 @@ public class ContractServiceImpl extends ContractRepository implements ContractS
   InvoiceLineService invoiceLineService = Beans.get(InvoiceLineService.class);
 
   public InvoiceLine generate(Invoice invoice, ContractLine line) throws AxelorException {
+
+    BigDecimal inTaxPriceComputed =
+        invoiceLineService.convertUnitPrice(false, line.getTaxLine(), line.getPrice());
     InvoiceLineGenerator invoiceLineGenerator =
         new InvoiceLineGenerator(
             invoice,
             line.getProduct(),
             line.getProductName(),
             line.getPrice(),
-            invoiceLineService.convertUnitPrice(false, line.getTaxLine(), line.getPrice()),
-            null,
+            inTaxPriceComputed,
+            invoice.getInAti() ? inTaxPriceComputed : line.getPrice(),
             line.getDescription(),
             line.getQty(),
             line.getUnit(),